[UPDATE] Modularize the print of the Stars status and change the Package name
