+++ step3/src/SixDegreesofKevinBacon/src/main/java/org/step3/Movie.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.step3;
 
 import java.util.ArrayList;
 import java.util.List;
